The authentication is already done, although some changes need to be done
