+++ gethired-frontend-web/components/Layout.jsx
@@ -1,11 +1,15 @@
 import React from "react"
 import Navbar from "./Navbar";
+import { useUser } from '@auth0/nextjs-auth0';
 
-const Layout = ({ userName, children }) => (
-  <React.Fragment>
-    <Navbar userName={userName} />
-    { children }
-  </React.Fragment>
-)
+const Layout = ({ userName, children }) => {
+  const {user} = useUser()
+  return(
+    <React.Fragment>
+      <Navbar userName={user.name} />
+      { children }
+    </React.Fragment>
+  )
+}
 
 export default Layout;
