+++ gethired-frontend-web/constants.js
@@ -1,1 +1,6 @@
 export const dateRegExp = new RegExp(/\d{4,4}-\d{2,2}-\d{2,2}/)
+
+export const options = [
+  { value: 'pie-chart', label: 'Pie Chart' },
+  { value: 'bar-chart', label: 'Bar Chart' },
+];
