+++ gethired-frontend-web/package.json
@@ -9,6 +9,7 @@
     "test": "jest --watch"
   },
   "dependencies": {
+    "@auth0/nextjs-auth0": "^1.7.0",
     "chart.js": "^3.6.2",
     "next": "^12.0.7",
     "react": "17.0.2",
