+++ gethired-frontend-web/pages/api/auth/[...auth0].js
@@ -1,0 +1,4 @@
+// pages/api/auth/[...auth0].js
+import { handleAuth } from '@auth0/nextjs-auth0';
+
+export default handleAuth();
